Add force calendar update option on gotoDate
When gotoDate method is called, it only updates the calendars when a new
calendar is introduced, by changing the month of the date. This causes a
side effect, when adding new calendars, the minDate not to update and
showing an empty calendar in some cases.

For this reason I added a force update option, in order to force the
update of the calendars in case we need it.